+++ src/main/java/mod/gottsch/forge/treasure2/core/event/WorldEventHandler.java
@@ -48,6 +48,7 @@ public class WorldEventHandler {
 
 	private static Path worldSavePath;
 	private static boolean isLoaded = false;
+	private static boolean isClientLoaded = false;
 
 	@SubscribeEvent(priority = EventPriority.HIGH)
 	public static void onWorldLoad(LevelEvent.Load event) {
@@ -89,6 +90,11 @@ public class WorldEventHandler {
 
 			} else {
 				Treasure.LOGGER.warn("unable to locate the world save folder.");
+			}
+		} else {
+			if (!isClientLoaded) {
+				TreasureDataFixer.fix();
+				isClientLoaded = true;
 			}
 		}
 	}