+++ src/main/java/mod/gottsch/forge/treasure2/core/item/KeyItem.java
@@ -70,7 +70,8 @@ import net.minecraftforge.common.util.LazyOptional;
  */
 public class KeyItem extends Item implements IKeyEffects {
 	public static final int DEFAULT_MAX_USES = 25;
-
+	public static final String DURABILITY_TAG = "treasure2:durability";
+	
 	/*
 	 * The category that the key belongs to
 	 */
@@ -134,22 +135,31 @@ public class KeyItem extends Item implements IKeyEffects {
 	 */
 	@Override
 	public CompoundTag getShareTag(ItemStack stack) {
-		CompoundTag tag = null;
+		super.getShareTag(stack);
+
+		CompoundTag capabilityTag = null;
 
 		IDurabilityHandler handler = stack.getCapability(DURABILITY).map(h -> h).orElse(null);
 		if (handler != null) {
-			tag = handler.save();
+			capabilityTag = handler.save();
 		}
-		return tag;
+		CompoundTag stackTag = stack.getOrCreateTag();
+		// NOTE must ensure to add the capability tag to the original stack tag.
+		if (capabilityTag != null) {
+			stackTag.put(DURABILITY_TAG, capabilityTag);
+		}
+		return stackTag;
 	}
 
 	@Override
 	public void readShareTag(ItemStack stack, @Nullable CompoundTag tag) {
 		super.readShareTag(stack, tag);
 
-		IDurabilityHandler handler = stack.getCapability(DURABILITY).map(h -> h).orElse(null);
-		if (handler != null) {
-			handler.load(tag);
+		if (tag.contains(DURABILITY_TAG)) {
+			IDurabilityHandler handler = stack.getCapability(DURABILITY).map(h -> h).orElse(null);
+			if (handler != null) {
+				handler.load(tag.get(DURABILITY_TAG));
+			}
 		}
 	}
 
@@ -178,7 +188,7 @@ public class KeyItem extends Item implements IKeyEffects {
 			});
 		}
 		else {
-			tooltip.add(Component.translatable(LangUtil.tooltip("durability.amount"), stack.getMaxDamage() - stack.getDamageValue(), stack.getMaxDamage()));
+			tooltip.add(Component.translatable(LangUtil.tooltip("cap.durability.amount"), stack.getMaxDamage() - stack.getDamageValue(), stack.getMaxDamage()));
 		}
 		
 		tooltip.add(Component.translatable(LangUtil.tooltip("key_lock.rarity"), ChatFormatting.BLUE + Component.translatable(getRarity().getValue().toLowerCase()).getString().toUpperCase() ));
@@ -314,7 +324,7 @@ public class KeyItem extends Item implements IKeyEffects {
 				
 				// check key's breakability
 				if (breakKey) {
-					if ((isBreakable() || anyLockBreaksKey(chestBlockEntity.getLockStates(), this)) && Config.SERVER.keysAndLocks.enableKeyBreaks.get()) {
+					if (!context.getPlayer().isCreative() && (isBreakable() || anyLockBreaksKey(chestBlockEntity.getLockStates(), this)) && Config.SERVER.keysAndLocks.enableKeyBreaks.get()) {
 
 						// this damage block is considering if a key has been merged with another key.
 						// it is only 'breaking' 1 key's worth of damage
@@ -342,7 +352,7 @@ public class KeyItem extends Item implements IKeyEffects {
 				}
 
 				// user attempted to use key - increment the damage
-				if (isDamageable(heldItemStack) && !isKeyBroken) {
+				if (!context.getPlayer().isCreative() && isDamageable(heldItemStack) && !isKeyBroken) {
 					heldItemStack.setDamageValue(heldItemStack.getDamageValue() + 1);
 					if (heldItemStack.getDamageValue() >= cap.getDurability()) {
 						heldItemStack.shrink(1);