+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.7.1] - 2024-01-25
+
+### Changed
+
+- Fixed key usage/damage not updating client when playing on dedicated server.
+- Key usage/damage isn't increment in Creative mode.
+
 ## [3.7.0] - 2024-01-17
 
 ### Added